x86, arm: Move JAILHOUSE_BORROW_ROOT_PT into jailhouse_header.h
JAILHOUSE_BORROW_ROOT_PT and JAILHOUSE_BASE belong together logically,
and they used to be prior to 646233947717. Restore this by moving the
former also into asm/jailhouse_header.h. This is the expected location
for it because JAILHOUSE_BORROW_ROOT_PT is only relevant during the
setup process that jailhouse_header.h is used for.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>